Keep the internal functions of spectrum_lightcurve.cc internal
exspec now calls write_partial_lightcurve_spectra(), so no other file uses the
other functions of spectrum_lightcurve.cc or the Spectra struct. They are now
in the anonymous namespace, and spectrum_lightcurve.h declares only
write_partial_lightcurve_spectra(). The functions are now in the order of use,
because the header no longer declares them.

The output files do not change.
